21.(목) 문자 메세지 형식 저장 및 배송상태 관리에서 문자 기능 추가

+++ SemiProject/src/util/paging/Pagination.java
@@ -142,7 +142,7 @@ public class Pagination {
 		this.totalCnt = totalCnt;
 		
 		// 전체 페이지 수
-		this.pageCnt = (int) Math.ceil(totalCnt/sizePerPage);
+		this.pageCnt = (int) Math.ceil((double)totalCnt/sizePerPage);
 		// 시작 페이지
 		this.startPage = (range - 1) * pageSize + 1;
 		// 끝 페이지
